fix a few warnings
Ignore-this: 8efbc7577310fc188d70fb65bde3193d

darcs-hash:20110110110645-2fc9d-1f6b66e81b1f76f428536ce408e75ba92ce13ad6.gz

+++ itkComponentTree.h
@@ -186,7 +186,7 @@ public:
   void NodeTakeIndexesFrom( NodeType * node, NodeType *obsoletedNode );
 
   /** Return the number of index  in the node and its children */
-  int NodeCountIndexes( const NodeType *node ) const;
+  unsigned long NodeCountIndexes( const NodeType *node ) const;
 
   const PixelType & GetPixel( const IndexType & idx ) const;
